% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyModel.R
\name{applyModel}
\alias{applyModel}
\title{applyModel}
\usage{
applyModel(model = NULL, newData = NULL)
}
\arguments{
\item{model}{fitted tidymodel workflow object}

\item{newData}{new data matching fitted data format}
}
\value{
prediction of new data using given model
}
\description{
applyModel
}
