% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCINStats.R
\name{calculateCINStats}
\alias{calculateCINStats}
\title{calculateCINStats}
\usage{
calculateCINStats(data = NULL)
}
\arguments{
\item{data}{data.frame or list of segmented copy number profiles}
}
\value{
data.frame
}
\description{
calculate CIN statistics and metrics over a set of segmented copy number
profiles
}
