% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateClonality.R
\name{calculateClonality}
\alias{calculateClonality}
\title{calculateClonality}
\usage{
calculateClonality(data = NULL, weighted = TRUE)
}
\arguments{
\item{data}{data.frame containing a segmented copy number profile}

\item{weighted}{boolean to use weighted.mean to compute clonality (default:
TRUE)}
}
\value{
numeric
}
\description{
Compute clonality (measure of segment noise) using either weighted or
non-weighted mean of absolute segment distance from integer
}
