% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHomozygousLoss.R
\name{calculateHomozygousLoss}
\alias{calculateHomozygousLoss}
\title{calculateHomozygousLoss}
\usage{
calculateHomozygousLoss(data = NULL, threshold = 0.2)
}
\arguments{
\item{threshold}{}
}
\description{
calculateHomozygousLoss
}
