% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePloidy.R
\name{calculatePloidy}
\alias{calculatePloidy}
\title{calculatePloidy}
\usage{
calculatePloidy(data)
}
\arguments{
\item{data}{data.frame containing a segmented copy number profile}
}
\value{
numeric
}
\description{
Compute ploidy using mean copy number segments, weighted by segment length.
}
