% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSegmentCounts.R
\name{calculateSegmentCounts}
\alias{calculateSegmentCounts}
\title{calculateSegmentCounts}
\usage{
calculateSegmentCounts(data = NULL)
}
\arguments{
\item{data}{copy number profile from which to count number of segments}
}
\value{
numeric
}
\description{
Compute the number of segments for a given copy number profile
}
