% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINfits-package.R
\docType{package}
\name{CINfits-package}
\alias{CINfits}
\alias{CINfits-package}
\title{CINfits: CINfits - Quality control and absolute fitting of copy number profiles}
\description{
CINfits is a package for performing quality control and copy number fitting on absolute copy number profiles. The package computes a range of useful metrics, interactive & non-interactive workflows, and allows for refitting of copy number profiles.
}
\author{
\strong{Maintainer}: Philip S Smith \email{philip.smith@cruk.cam.ac.uk} (\href{https://orcid.org/0000-0002-9306-1747}{ORCID}) [copyright holder]

}
\keyword{internal}
