% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLMLogisticRegression.R
\name{fitGLMLogisticRegression}
\alias{fitGLMLogisticRegression}
\title{fitGLMLogisticRegression}
\usage{
fitGLMLogisticRegression(
  data = NULL,
  model = NULL,
  folds = NULL,
  mixture = 1,
  metric = "accuracy"
)
}
\arguments{
\item{data}{model data}

\item{model}{model recipe object}

\item{folds}{crossfold validation object}

\item{mixture}{GLM lasso-ridge regression mixture (Default: 1).
Value of 0 is ridge, 1 is lasso, float values between are elastic net.}

\item{metric}{model performance metric (Default: "accuracy")}
}
\value{
fitted model
}
\description{
fitGLMLogisticRegression
}
