% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitInteractive.R
\name{fitInteractive}
\alias{fitInteractive}
\title{fitInteractive}
\usage{
fitInteractive(data = NULL, metadata = NULL)
}
\arguments{
\item{data}{data.frame or list of segmented copy number profiles}

\item{metadata}{data.frame containing meta data for the samples contained in
data. This should include at least 3 columns; 'sample', 'ploidy', and
'purity'.}
}
\value{
interactive shiny application
}
\description{
Launch an interactive shiny application  to perform copy number fitting and
quality control
}
