% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRandomForest.R
\name{fitRandomForest}
\alias{fitRandomForest}
\title{fitRandomForest}
\usage{
fitRandomForest(
  data = NULL,
  model = NULL,
  folds = NULL,
  metric = "accuracy",
  importance = "impurity",
  trees = 1000
)
}
\arguments{
\item{data}{model data}

\item{model}{model recipe object}

\item{folds}{crossfold validation object}

\item{metric}{model performance metric (Default: "accuracy")}

\item{importance}{Ranger RF engine importance parameter (Default: "impurity")}

\item{trees}{number of trees (Default: 1000)}
}
\value{
fitted model
}
\description{
fitRandomForest
}
