% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSVM.R
\name{fitSVM}
\alias{fitSVM}
\title{fitSVM}
\usage{
fitSVM(data = NULL, model = NULL, folds = NULL, metric = "accuracy")
}
\arguments{
\item{data}{model data}

\item{model}{model recipe object}

\item{folds}{crossfold validation object}

\item{metric}{model performance metric (Default: "accuracy")}
}
\value{
fitted model
}
\description{
fitSVM
}
