% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitXGBtree.R
\name{fitXGBtree}
\alias{fitXGBtree}
\title{fitXGBtree}
\usage{
fitXGBtree(
  data = NULL,
  model = NULL,
  folds = NULL,
  metric = "accuracy",
  trees = 1000
)
}
\arguments{
\item{data}{model data}

\item{model}{model recipe object}

\item{folds}{crossfold validation object}

\item{metric}{model performance metric (Default: "accuracy")}

\item{trees}{number of trees (Default: 1000)}
}
\value{
fitted model
}
\description{
fitXGBtree
}
