% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContingencyTable.R
\name{getContingencyTable}
\alias{getContingencyTable}
\title{getContingencyTable}
\usage{
getContingencyTable(model = NULL, var = "use")
}
\arguments{
\item{model}{fitted tidymodel workflow object}

\item{var}{class label variable name (default: use)}
}
\value{
contingency table
}
\description{
getContingencyTable
}
