% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModelRecipe.R
\name{makeModelRecipe}
\alias{makeModelRecipe}
\title{makeModelRecipe}
\usage{
makeModelRecipe(data = NULL, folds = 10, strata = "use")
}
\arguments{
\item{data}{training data}

\item{folds}{number of folds to use in cross validation}

\item{strata}{variable on which to stratify training/test data for class balance}
}
\value{
recipe object for tidymodels
}
\description{
makeModelRecipe
}
