% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSunrise.R
\name{plotSunrise}
\alias{plotSunrise}
\title{plotSunrise}
\usage{
plotSunrise(data = NULL, ploidys = NULL, purities = NULL)
}
\arguments{
\item{data}{data.frame containing a segmented copy number profile}

\item{ploidys}{numeric vector of ploidy values (Default:
"seq.int(1.2,8,0.1)")}

\item{purities}{numeric vector of purity values between 0 and 1.0 (Default:
"seq.int(0.2,1,0.01)")}
}
\value{
plot
}
\description{
Plot sunrise plot of clonality values computed across a gridsearch of
supplied ploidy and purity values
}
