% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAlleleSpecific.R
\name{processAlleleSpecific}
\alias{processAlleleSpecific}
\title{processAlleleSpecific}
\usage{
processAlleleSpecific(
  data = NULL,
  nameMaj = "nAraw",
  nameMinor = "nBraw",
  chrName = "chr",
  startName = "startpos",
  endName = "endpos",
  sampleName = "sample",
  returnTotal = TRUE,
  returnSmoothed = FALSE,
  smoothingFactor = 0.1
)
}
\arguments{
\item{data}{absolute copy nubmer profile containing allele specific copy states.}

\item{nameMaj}{column name of major allele column (default: nAraw).}

\item{nameMinor}{column name of minor allele column (default: nBraw).}

\item{chrName}{column name of chromosome column (default: chr).}

\item{startName}{column name of genome start position column (default: startpos).}

\item{endName}{column name of genome end position column (default: endpos).}

\item{sampleName}{column name of sample identifier column (default: sample).}

\item{returnTotal}{A boolean value as to whether to additionally return total copy number, returned as 'segVal' (default: TRUE).}

\item{returnSmoothed}{A boolean value as to whether to return smoothed total copy number and drop allele-specific information.}

\item{smoothingFactor}{A float between 0 and 1 which is used to smooth and collapse adjacent total copy number segments (default: 0.1).}
}
\value{
data.frame
}
\description{
processAlleleSpecific
}
