% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothProfile.R
\name{smoothProfile}
\alias{smoothProfile}
\title{smoothProfile}
\usage{
smoothProfile(
  data = NULL,
  smoothingFactor = 0.12,
  implementation = "linear",
  method = "median",
  alleleSpecific = FALSE
)
}
\arguments{
\item{data}{data.frame containing copy number segments}

\item{smoothingFactor}{Distance between two segments at which to collapse
into a single segment (Default: 0.12)}

\item{implementation}{Smoothing implementation to use, either linear or
iterative (Default: linear)}

\item{method}{method to collapse segments (Default: median). Options are
median, mean, weighted.mean.}

\item{alleleSpecific}{use allele-specific implementation}
}
\value{
data.frame
}
\description{
smoothProfile
}
