% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFittingData.R
\name{splitFittingData}
\alias{splitFittingData}
\title{splitFittingData
Split manually fitted QC data into training and test data}
\usage{
splitFittingData(data = NULL, prop = 0.7, strata = "use")
}
\arguments{
\item{data}{QC data table}

\item{prop}{proportion to retain for training}

\item{strata}{variable to stratify in order to maintain class balance}
}
\value{
list of sample split, training split, and testing split samples
}
\description{
splitFittingData
Split manually fitted QC data into training and test data
}
