% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitInteractive.R
\name{fitInteractive}
\alias{fitInteractive}
\title{fitInteractive}
\usage{
fitInteractive(data = NULL)
}
\arguments{
\item{data}{data.frame or list of segmented copy number profiles}
}
\value{
interactive shiny application
}
\description{
Launch an interactive shiny application  to perform copy number fitting and
quality control
}
