% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plotProfile}
\alias{plotProfile}
\title{plotProfile}
\usage{
plotProfile(
  data = NULL,
  sample = NULL,
  cn.max = 15,
  purity = NULL,
  alleleSpecific = FALSE,
  cols = NULL
)
}
\arguments{
\item{data}{data.frame containing a segmented copy number profile}

\item{sample}{vector of length 1 containing either a sample name or sample index}

\item{cn.max}{maximum copy number to plot - Values greater than this are truncated to the specified value}

\item{purity}{given purity for profile being plotted - not required for non-interactive use.}

\item{alleleSpecific}{Boolean as to whether plot total or allele-specific copy number.}

\item{cols}{colours for segments plotted on copy number profile}
}
\value{
plot
}
\description{
Produce a plot of segments from a given segmented copy number profile
}
