% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleFit.R
\name{rescaleFit}
\alias{rescaleFit}
\title{rescaleFit}
\usage{
rescaleFit(data = NULL, ploidy = NULL, purity = NULL, segValOnly = FALSE)
}
\arguments{
\item{data}{data.frame containing a segmented copy number profile}

\item{ploidy}{new ploidy value to refit copy number profile}

\item{purity}{new ploidy value to refit copy number profile}

\item{segValOnly}{return only segVal vector rather than entire refitted segment
table (Default: FALSE)}
}
\value{
data.frame
}
\description{
Refit a segmented copy number profile to a new ploidy-purity combination
}
